# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/migrations": {
            "post": {
                "description": "Migrations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Migrations"
                ],
                "summary": "Migrations",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Migration"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "ping",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ping"
                ],
                "summary": "ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{id}/transactions": {
            "get": {
                "description": "Transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "from",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "to",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.TransactionResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Migration": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "csv_path": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "processed_lines": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/domain.MigrationStatus"
                },
                "summary": {
                    "type": "string"
                },
                "total_lines": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.MigrationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "complete",
                "failed",
                "processing"
            ],
            "x-enum-varnames": [
                "Pending",
                "Complete",
                "Failed",
                "Processing"
            ]
        },
        "domain.TransactionResult": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number",
                    "default": 0
                },
                "total_credits": {
                    "type": "number",
                    "default": 0
                },
                "total_debits": {
                    "type": "number",
                    "default": 0
                }
            }
        }
    }
}